Use Zcpy API code path for non-RDMA layers
