+++ src/ck-core/ck.h
@@ -28,12 +28,10 @@ inline void _CldEnqueue(int pe, void *msg, int infofn) {
     CmiFree(msg);
     return;
   }
-#if CMK_ONESIDED_IMPL
   envelope *env = (envelope *)msg;
   // Store source information to handle acknowledgements on completion
   if(CMI_IS_ZC(msg))
     CkRdmaPrepareZCMsg(env, CkNodeOf(pe));
-#endif
   CldEnqueue(pe, msg, infofn);
 }
 inline void _CldEnqueueWithinNode(void *msg, int infofn) {
@@ -62,33 +60,27 @@ inline void _CldNodeEnqueue(int node, void *msg, int infofn) {
     CmiFree(msg);
     return;
   }
-#if CMK_ONESIDED_IMPL
   envelope *env = (envelope *)msg;
   // Store source information to handle acknowledgements on completion
   if(CMI_IS_ZC(msg))
     CkRdmaPrepareZCMsg(env, node);
-#endif
   CldNodeEnqueue(node, msg, infofn);
 }
 #else
 
 inline void _CldEnqueue(int pe, void *msg, int infofn) {
-#if CMK_ONESIDED_IMPL
   envelope *env = (envelope *)msg;
   // Store source information to handle acknowledgements on completion
   if(CMI_IS_ZC(msg))
     CkRdmaPrepareZCMsg(env, CkNodeOf(pe));
-#endif
   CldEnqueue(pe, msg, infofn);
 }
 
 inline void _CldNodeEnqueue(int node, void *msg, int infofn) {
-#if CMK_ONESIDED_IMPL
   envelope *env = (envelope *)msg;
   // Store source information to handle acknowledgements on completion
   if(CMI_IS_ZC(msg))
     CkRdmaPrepareZCMsg(env, node);
-#endif
   CldNodeEnqueue(node, msg, infofn);
 }
 #define _CldEnqueueMulti      CldEnqueueMulti
