+++ src/ck-core/ckrdma.C
@@ -11,6 +11,8 @@
 #include "ckarray.h"
 #include "cklocation.h"
 
+void CmiFreeBroadcastAllExceptMeFn(int size, char *msg);
+
 #if CMK_SMP
 /*readonly*/ extern CProxy_ckcallback_group _ckcallbackgroup;
 #endif
@@ -290,7 +292,7 @@ void CkRdmaDirectAckHandler(void *ack) {
   switch(info->opMode) {
     case CMK_DIRECT_API             : handleDirectApiCompletion(info); // Ncpy Direct API
                                       break;
-#if CMK_ONESIDED_IMPL
+
     case CMK_EM_API                 : handleEntryMethodApiCompletion(info); // Ncpy EM API invoked through a GET
                                       break;
 
@@ -311,7 +313,7 @@ void CkRdmaDirectAckHandler(void *ack) {
 
     case CMK_READONLY_BCAST         : readonlyGetCompleted(info);
                                       break;
-#endif
+
     case CMK_ZC_PUP                 : zcPupGetCompleted(info);
                                       break;
 
@@ -373,7 +375,6 @@ inline void zcQdIncrement() {
 }
 
 
-#if CMK_ONESIDED_IMPL
 /*********************************** Zerocopy Entry Method API ****************************/
 
 /************************* Zerocopy Entry Method API - Utility functions ******************/
@@ -1599,7 +1600,7 @@ void CkReplaceSourcePtrsInBcastMsg(envelope *env, NcpyBcastInterimAckInfo *bcast
 #endif
 }
 
-#if CMK_SMP
+#if CMK_SMP || CMK_UTH_VERSION
 void updatePeerCounterAndPush(envelope *env) {
   int pe;
   int numops, rootNode;
@@ -1693,7 +1694,7 @@ void handleArrayMsgOnChildPostCompletionForRecvBcast(envelope *env) {
   CkArray *mgr = getArrayMgrFromMsg(env);
   mgr->forwardZCMsgToOtherElems(env);
 
-#if CMK_SMP
+#if CMK_SMP || CMK_UTH_VERSION
   if(CmiMyNodeSize() > 1) {
     sendRecvDoneMsgToPeers(env, mgr);
   } else
@@ -1707,7 +1708,7 @@ void handleArrayMsgOnChildPostCompletionForRecvBcast(envelope *env) {
 
 void handleGroupMsgOnChildPostCompletionForRecvBcast(envelope *env) {
   CMI_ZC_MSGTYPE(env) = CMK_ZC_BCAST_RECV_DONE_MSG;
-#if CMK_SMP
+#if CMK_SMP || CMK_UTH_VERSION
   if(CmiMyNodeSize() > 1) {
     sendRecvDoneMsgToPeers(env, NULL);
   } else
@@ -1915,7 +1916,8 @@ void readonlyGetCompleted(NcpyOperationInfo *ncpyOpInfo) {
   }
 
   // Free ncpyOpInfo allocated inside readonlyGet
-  CmiFree(ncpyOpInfo);
+  if(ncpyOpInfo->freeMe == CMK_FREE_NCPYOPINFO)
+    CmiFree(ncpyOpInfo);
 }
 /***************************** End of Zerocopy Readonly Bcast Support ****************************/
 
@@ -1958,14 +1960,11 @@ inline void _ncpyBcastNoHandler(ncpyBcastNoMsg *bcastNoMsg) {
   arrProxy[bcastNoMsg->srcPe].sendZCBroadcast(w);
 }
 
-#endif  /* End of CMK_ONESIDED_IMPL */
 
 // Register converse handler for invoking ncpy ack
 void initEMNcpyAckHandler(void) {
-#if CMK_ONESIDED_IMPL
   ncpy_bcastNo_handler_idx = CmiRegisterHandler((CmiHandler)_ncpyBcastNoHandler);
   ncpy_handler_idx = CmiRegisterHandler((CmiHandler)_ncpyAckHandler);
-#endif
 #if CMK_SMP
   zcpy_pup_complete_handler_idx = CmiRegisterHandler((CmiHandler)_zcpyPupCompleteHandler);
 #endif
@@ -1981,7 +1980,6 @@ inline void invokeRemoteNcpyAckHandler(int pe, void *ref, ncpyHandlerIdx opMode)
   CmiSyncSendAndFree(pe, sizeof(ncpyHandlerMsg), (char *)msg);
 }
 
-#if CMK_ONESIDED_IMPL
 inline void _ncpyAckHandler(ncpyHandlerMsg *msg) {
   QdProcess(1);
 
@@ -2002,7 +2000,6 @@ inline void _ncpyAckHandler(ncpyHandlerMsg *msg) {
                                                  break;
   }
 }
-#endif  /* End of CMK_ONESIDED_IMPL */
 
 
 /***************************** Zerocopy PUP Support ****************************/
