+++ src/ck-core/init.C
@@ -123,7 +123,6 @@ UInt  _numInitMsgs = 0;
  */
 CksvDeclare(UInt,_numInitNodeMsgs);
 
-#if CMK_ONESIDED_IMPL
 #if CMK_SMP
 std::atomic<UInt> numZerocopyROops = {0};
 #else
@@ -134,7 +133,6 @@ bool usedCMAForROBcastTransfer = false;
 NcpyROBcastAckInfo *roBcastAckInfo;
 int   _roRdmaDoneHandlerIdx;
 CksvDeclare(int,  _numPendingRORdmaTransfers);
-#endif
 
 int   _infoIdx;
 int   _charmHandlerIdx;
@@ -905,7 +903,6 @@ static inline void _processRODataMsg(envelope *env)
     PUP::fromMem pu((char *)EnvToUsr(env));
     CmiSpanningTreeInfo &t = *_topoTree;
 
-#if CMK_ONESIDED_IMPL
 #if CMK_SMP
     UInt numZerocopyROopsTemp;
     pu|numZerocopyROopsTemp;
@@ -921,14 +918,13 @@ static inline void _processRODataMsg(envelope *env)
     if(numZerocopyROops > 0 && t.child_count != 0) {
       readonlyAllocateOnSource();
     }
-#endif
 
     for(size_t i=0;i<_readonlyTable.size();i++) {
       _readonlyTable[i]->pupData(pu);
     }
 
     // Forward message to peers after numZerocopyROops has been initialized
-#if CMK_ONESIDED_IMPL && CMK_SMP
+#if CMK_SMP
     if(CMI_IS_ZC_BCAST(env)) {
       // Send message to peers
       CmiForwardMsgToPeers(env->getTotalsize(), (char *)env);
@@ -961,7 +957,6 @@ static void _roRestartHandler(void *msg)
   _triggerHandler(NULL);
 }
 
-#if CMK_ONESIDED_IMPL
 static void _roRdmaDoneHandler(envelope *env) {
 
   switch(env->getMsgtype()) {
@@ -1005,12 +1000,10 @@ static void _roRdmaDoneHandler(envelope *env) {
       break;
   }
 }
-#endif
 
 void checkForInitDone(bool rdmaROCompleted) {
 
   bool noPendingRORdmaTransfers = true;
-#if CMK_ONESIDED_IMPL
   // Ensure that there are no pending RO Rdma transfers on Rank 0 when numZerocopyROops > 0
 #if CMK_SMP
   if(numZerocopyROops.load(std::memory_order_relaxed) > 0)
@@ -1018,7 +1011,6 @@ void checkForInitDone(bool rdmaROCompleted) {
   if(numZerocopyROops > 0)
 #endif
     noPendingRORdmaTransfers = rdmaROCompleted;
-#endif
   if (_numExpectInitMsgs && CkpvAccess(_numInitsRecd) + CksvAccess(_numInitNodeMsgs) == _numExpectInitMsgs && noPendingRORdmaTransfers)
     _initDone();
 }
@@ -1082,7 +1074,7 @@ static void _initHandler(void *msg, CkCoreState *ck)
       CmiAbort("Internal Error: Unknown-msg-type. Contact Developers.\n");
   }
   DEBUGF(("[%d,%.6lf] _numExpectInitMsgs %d CkpvAccess(_numInitsRecd)+CksvAccess(_numInitNodeMsgs) %d+%d\n",CmiMyPe(),CmiWallTimer(),_numExpectInitMsgs,CkpvAccess(_numInitsRecd),CksvAccess(_numInitNodeMsgs)));
-#if CMK_ONESIDED_IMPL && CMK_USE_CMA
+#if CMK_USE_CMA
   if(usedCMAForROBcastTransfer) {
     checkForInitDone(true); // RO ZCPY Bcast operation was completed inline (using CMA)
   } else
@@ -1250,7 +1242,6 @@ void _sendReadonlies() {
   //Determine the size of the RODataMessage
   PUP::sizer ps;
 
-#if CMK_ONESIDED_IMPL
 #if CMK_SMP
   UInt numZerocopyROopsTemp;
   numZerocopyROopsTemp = numZerocopyROops;
@@ -1258,26 +1249,21 @@ void _sendReadonlies() {
 #else
   ps|numZerocopyROops;
 #endif
-#endif
 
   for(int i=0;i<_readonlyTable.size();i++) _readonlyTable[i]->pupData(ps);
 
-#if CMK_ONESIDED_IMPL
   if(CkNumNodes() > 1 && numZerocopyROops > 0) { // No ZC ops are performed when CkNumNodes <= 1
     readonlyAllocateOnSource();
   }
-#endif
 
   //Allocate and fill out the RODataMessage
   envelope *env = _allocEnv(RODataMsg, ps.size());
   PUP::toMem pp((char *)EnvToUsr(env));
-#if CMK_ONESIDED_IMPL
 #if CMK_SMP
   numZerocopyROopsTemp = numZerocopyROops;
   pp|numZerocopyROopsTemp;
 #else
   pp|numZerocopyROops;
-#endif
 #endif
   for(int i=0;i<_readonlyTable.size();i++) _readonlyTable[i]->pupData(pp);
 
@@ -1286,7 +1272,7 @@ void _sendReadonlies() {
   CmiSetHandler(env, _initHandlerIdx);
   DEBUGF(("[%d,%.6lf] RODataMsg being sent of size %d \n",CmiMyPe(),CmiWallTimer(),env->getTotalsize()));
   CmiSyncBroadcast(env->getTotalsize(), (char *)env);
-#if CMK_ONESIDED_IMPL && CMK_SMP
+#if CMK_SMP
   if(numZerocopyROops > 0) {
     // Send message to peers
     CmiForwardMsgToPeers(env->getTotalsize(), (char *)env);
@@ -1357,9 +1343,7 @@ void _initCharm(int unused_argc, char **argv)
 
 	CksvInitialize(objNumRdmaOpsMap, pendingZCOps);
 
-#if CMK_ONESIDED_IMPL
 	CksvInitialize(int, _numPendingRORdmaTransfers);
-#endif
 
 	CkpvInitialize(_CkOutStream*, _ckout);
 	CkpvInitialize(_CkErrStream*, _ckerr);
@@ -1380,9 +1364,7 @@ void _initCharm(int unused_argc, char **argv)
 	  	CksvAccess(_numNodeGroups) = 1; //make 0 an invalid group number
           	CksvAccess(_numInitNodeMsgs) = 0;
 
-#if CMK_ONESIDED_IMPL
 		CksvAccess(_numPendingRORdmaTransfers) = 0;
-#endif
 
 		CksvAccess(_nodeLock) = CmiCreateLock();
 		CksvAccess(_nodeGroupTable) = new GroupTable();
@@ -1413,9 +1395,7 @@ void _initCharm(int unused_argc, char **argv)
 	CmiAssignOnce(&_initHandlerIdx, CkRegisterHandlerEx(_initHandler, CkpvAccess(_coreState)));
 	CmiAssignOnce(&_roRestartHandlerIdx, CkRegisterHandler(_roRestartHandler));
 
-#if CMK_ONESIDED_IMPL
 	CmiAssignOnce(&_roRdmaDoneHandlerIdx, CkRegisterHandler(_roRdmaDoneHandler));
-#endif
 
 	CmiAssignOnce(&_exitHandlerIdx, CkRegisterHandler(_exitHandler));
 	//added for interoperabilitY
